requirements.txt
# Paquetes requeridos para la aplicación Divisor de Gastos de Restaurante

streamlit
streamlit
google-generativeai
Pillow
pandas
python-dotenv